requirements.txt
Django==3.2
python-dotenv==0.21.0
sorl-thumbnail==12.9.0
Pillow==9.3.0
django-cleanup==6.0.0